requirements.txt
torch
rdkit-pypi
fastapi
uvicorn
pandas
scikit-learn
tqdm
